+++ middleware/timeout/README.md
@@ -1,50 +1,23 @@
 # Timeout
-Timeout middleware for [Fiber](https://github.com/gofiber/fiber) wraps a `fiber.Handler` with a timeout. If the handler takes longer than the given duration to return, the timeout error is set and forwarded to the centralized [ErrorHandler](https://docs.gofiber.io/error-handling).
+Timeout middleware for Fiber. As a `fiber.Handler` wrapper, it creates a context with `context.WithTimeout` and pass it in `UserContext`.
 
-Also has timeout with context middleware by `NewWithContext`, it creates a context with `context.WithTimeout` and pass it in `UserContext`.
+If the context passed executions (eg. DB ops, Http calls) takes longer than the given duration to return, the timeout error is set and forwarded to the centralized `ErrorHandler`.
 
-If the context passed executions (eg. DB ops, Http calls) takes longer than the given duration to return, it cancels current operations using context and  the timeout error is set and forwarded to the centralized `ErrorHandler`. `UserContext` needs to be passed all long running operations.
-
-`NewWithContext` has no race conditions, ready to use on production.
-
+It has no race conditions, ready to use on production.
 
 ### Table of Contents
 - [Timeout](#timeout)
 		- [Table of Contents](#table-of-contents)
 		- [Signatures](#signatures)
 		- [Examples](#examples)
-		- [NewWithContext Examples](#newwithcontext-examples)
 
 
 ### Signatures
 ```go
-func New(h fiber.Handler, t time.Duration) fiber.Handler
-func NewWithContext(h fiber.Handler, t time.Duration, tErrs ...error) fiber.Handler
+func New(handler fiber.Handler, timeout time.Duration, timeoutErrors ...error) fiber.Handler
 ```
 
 ### Examples
-Import the middleware package that is part of the Fiber web framework
-```go
-import (
-  "github.com/gofiber/fiber/v2"
-  "github.com/gofiber/fiber/v2/middleware/timeout"
-)
-```
-
-After you initiate your Fiber app, you can use the following possibilities:
-```go
-handler := func(ctx *fiber.Ctx) error {
-	err := ctx.SendString("Hello, World 👋!")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-app.Get("/foo", timeout.New(handler, 5 * time.Second))
-```
-
-### NewWithContext Examples
 Import the middleware package that is part of the Fiber web framework
 ```go
 import (
@@ -63,7 +36,7 @@ h := func(c *fiber.Ctx) error {
 		return nil
 	}
 
-app.Get("/foo", timeoutcontext.NewWithContext(h, 5 * time.Second))
+app.Get("/foo", timeoutcontext.New(h, 5 * time.Second))
 
 func sleepWithContext(ctx context.Context, d time.Duration) error {
 	timer := time.NewTimer(d)
@@ -77,7 +50,6 @@ func sleepWithContext(ctx context.Context, d time.Duration) error {
 	}
 	return nil
 }
-
 ```
 
 Use with custom error:
@@ -92,7 +64,7 @@ h := func(c *fiber.Ctx) error {
 		return nil
 	}
 
-app.Get("/foo", timeoutcontext.NewWithContext(h, 5 * time.Second), ErrFooTimeOut)
+app.Get("/foo", timeoutcontext.New(h, 5 * time.Second), ErrFooTimeOut)
 
 func sleepWithContext(ctx context.Context, d time.Duration) error {
 	timer := time.NewTimer(d)
@@ -106,5 +78,4 @@ func sleepWithContext(ctx context.Context, d time.Duration) error {
 	}
 	return nil
 }
-
 ```
